congifuration using json file

+++ server/server.js
@@ -38,14 +38,15 @@ app.get('/todos',authenticate, (req, res) => {
   });
 });
 
-app.get('/todos/:id', (req, res) => {
+app.get('/todos/:id',authenticate, (req, res) => {
   var id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
   }
-
-  Todo.findById(id).then((todo) => {
+  Todo.findById({_id:id,
+    _creator:req.user._id})
+    .then((todo) => {
     if (!todo) {
       return res.status(404).send();
     }
@@ -56,14 +57,14 @@ app.get('/todos/:id', (req, res) => {
   });
 });
 
-app.delete('/todos/:id', (req, res) => {
+app.delete('/todos/:id',authenticate, (req, res) => {
   var id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
   }
 
-  Todo.findByIdAndRemove(id).then((todo) => {
+  Todo.findByIdAndRemove({_id:id,_creator:req.user._id}).then((todo) => {
     if (!todo) {
       return res.status(404).send();
     }
